fix: 유저 정보 조회 & 삭제 수정
