+++ src/main/java/com/sparta/newsfeed_project/domain/buddies/service/BuddiesService.java
@@ -28,8 +28,8 @@ public class BuddiesService {
      * @since 2024-10-23
      */
     public ResponseBuddiesDto createBuddies(Long memberId, Long userId) throws ResponseException {
-        if(!memberRepository.existsById(userId)) {
-            throw new ResponseException(ResponseCode. MEMBER_NOT_FOUND);
+        if (!memberRepository.existsById(userId)) {
+            throw new ResponseException(ResponseCode.MEMBER_NOT_FOUND);
         }
         if (memberId.equals(userId)) {
             throw new ResponseException(ResponseCode.BUDDIES_DUOLICATION_ID_ERROR);
@@ -52,7 +52,7 @@ public class BuddiesService {
      */
     public List<ResponseBuddiesDto> getBuddies() throws ResponseException {
         List<Buddies> buddies = buddiesRepository.findAll();
-        if(buddies.isEmpty()) {
+        if (buddies.isEmpty()) {
             throw new ResponseException(ResponseCode.BUDDIES_NULL_ERROR);
         }
         return buddies.stream().map(Buddies::to).toList();
@@ -65,7 +65,7 @@ public class BuddiesService {
      * @since 2024-10-23
      */
     public void acceptBuddies(Long memberId, Long userId) throws ResponseException {
-        if(!buddiesRepository.existsById(userId)) {
+        if (!buddiesRepository.existsById(userId)) {
             throw new ResponseException(ResponseCode.BUDDIES_DUOLICATION_ID_ERROR);
         }
         Buddies buddies = buddiesRepository
@@ -84,14 +84,14 @@ public class BuddiesService {
     public void deleteBuddies(Long memberId, Long userId) {
         Buddies buddies = buddiesRepository
                 .findOneByFromUserIdAndToUserId(memberId, userId);
-        if(buddies==null) {
+        if (buddies == null) {
             throw new ResponseException(ResponseCode.BUDDIES_NULL_ERROR);
         }
         buddies.approved(false);
 
         Buddies buddy = buddiesRepository
                 .findOneByFromUserIdAndToUserId(userId, memberId);
-        if(buddy==null) {
+        if (buddy == null) {
             throw new ResponseException(ResponseCode.BUDDIES_NULL_ERROR);
         }
 
@@ -102,6 +102,7 @@ public class BuddiesService {
             buddiesRepository.save(buddies);
         }
     }
+
     /**
      * 회원의 친구 관계를 모두 불러옵니다
      *
@@ -110,46 +111,10 @@ public class BuddiesService {
      */
     public List<ResponseBuddiesDto> getAllBuddies(Long memberId) {
         List<Buddies> buddies = buddiesRepository.findAllByToUserId(memberId);
-        if(buddies.isEmpty()) {
+        if (buddies.isEmpty()) {
             throw new ResponseException(ResponseCode.BUDDIES_NULL_ERROR);
         }
         return buddies.stream().map(Buddies::to).toList();
-    }
-
-
-
-
-    /**
-     * 회원의 친구 관계를 모두 삭제합니다.
-     *
-     * @param memberId 회원 ID
-     * @since 2024-10-23
-     */
-    public void deleteMemberBuddies(Long memberId) {
-        List<Buddies> buddies = buddiesRepository.findByFromUserIdOrToUserId(memberId, memberId);
-        if (buddies != null && !buddies.isEmpty())
-            buddiesRepository.deleteAll(buddies);
-
-    }
-
-    /**
-     * 회원이 친구 추가한 친구 수를 조회합니다.
-     *
-     * @param memberId 회원 ID
-     * @return 친구 수
-     */
-    public int getFromBuddyCount(Long memberId) {
-        return buddiesRepository.countByFromUserId(memberId).intValue();
-    }
-
-    /**
-     * 회원을 친구 추가한 친구 수를 조회합니다.
-     *
-     * @param memberId 회원 ID
-     * @return 친구 수
-     */
-    public int getToBuddyCount(Long memberId) {
-        return buddiesRepository.countByToUserId(memberId).intValue();
     }
 }
